elastic_streamlit/requirements.txt
streamlit==1.4.0
elasticsearch==7.17.0
requests==2.31.0
uvicorn==0.18.3
fastapi==0.85.1
transformers==4.25.1
torch==1.12.1
numpy==1.23.5
protobuf==3.20.3
altair==4.1.0
openai==1.6.1
click==7.1.2
python-dotenv 
langchain==0.0.123
langchain-openai==0.0.2
httpx==0.27.2
google.generativeai 
asyncio
PyMuPDF